change type of offer price

+++ src/main/java/com/example/gymcenterapp/entities/Offer.java
@@ -34,7 +34,7 @@ public class Offer
     private Integer offerPeriod;
 
     @Column(name = "offer_price")
-    private Integer offerPrice;
+    private Float offerPrice;
 
     @ManyToOne(cascade = { CascadeType.MERGE, CascadeType.REMOVE })
     @JoinColumn(name = "offer_activity")
